+++ src/main/java/pro/sky/calculatorproject/calculator/service/CalculatorService.java
@@ -3,12 +3,12 @@ package pro.sky.calculatorproject.calculator.service;
 import org.springframework.web.bind.annotation.GetMapping;
 
 public interface CalculatorService {
-    double getSum(double num1, double num2);
+    Integer getSum(Integer num1, Integer num2);
 
-    double getSubtraction(double num1, double num2);
+    Integer getSubtraction(Integer num1, Integer num2);
 
-    double getMultiply(double num1, double num2);
+    Integer getMultiply(Integer num1, Integer num2);
 
-    double getDivision(double num1, double num2);
+    Double getDivision(Double num1, Double num2);
 
 }
